test: add CSP integration spec wired against current sanitizer
Iterates the malicious-payload corpus and clean baselines through the
production sanitization pipeline and feeds the result into the CSP
replica harness. Baseline run: 31 failed / 57 passed. The MS-cert
content-url payload fails as expected; the img-src variant is already
covered by the existing src-attribute data-URI transformer. Full
baseline output captured in harness-baseline.log for the PR description.

Adds test:integration / test:all / cert-check npm scripts and a JSDOM
setup module so the spec can import the sanitization pipeline cleanly
under Playwright's Node test environment. The sanitization functions
are extracted from src/domain-utils.ts into a new src/sanitize-pipeline.ts
module so the harness can import them without dragging in d3-selection,
overlayscrollbars, or powerbi-visuals-api at test load time. domain-utils
re-exports getParsedHtmlAsDom for backwards compatibility with the visual
code and existing unit tests (184/184 still passing).

Switches test-integration/tsconfig.json to module: commonjs (Playwright
runs ts-node in CJS mode by default) and updates runner.ts to use the
CJS __dirname builtin.
