+++ package.json
@@ -12,7 +12,10 @@
         "test": "vitest run",
         "test:watch": "vitest",
         "test:ui": "vitest --ui",
-        "test:coverage": "vitest run --coverage"
+        "test:coverage": "vitest run --coverage",
+        "test:integration": "playwright test --config=test-integration/playwright.config.ts",
+        "test:all": "npm run test && npm run test:integration",
+        "cert-check": "npm run package && npm run test:integration"
     },
     "devDependencies": {
         "@playwright/test": "^1.59.1",
